Fix compiler warnings

+++ src/dividing_cell_op.h
@@ -15,7 +15,7 @@ class DividingCellOp {
   template <typename TContainer>
   void operator()(TContainer* cells, uint16_t type_idx) const {
 #pragma omp parallel for
-    for (size_t i = 0; i < cells->size(); i++) {
+    for (uint64_t i = 0; i < cells->size(); i++) {
       auto&& cell = (*cells)[i];
       if (cell.GetDiameter() <= 40) {
         cell.ChangeVolume(300);
